requirements.txt
asgiref==3.8.1
coverage==7.10.7
Django==5.2
factory_boy==3.3.3
Faker==37.8.0
flake8==7.3.0
inflection==0.5.1
iniconfig==2.1.0
mccabe==0.7.0
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-factoryboy==2.8.1
python-dotenv==1.1.0
ruff==0.13.2
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
